+++ src/imio/smartweb/core/contents/sections/contact/view.py
@@ -5,7 +5,7 @@ from datetime import timedelta
 from decimal import Decimal
 from imio.smartweb.common.utils import rich_description
 from imio.smartweb.core.config import DIRECTORY_URL
-from imio.smartweb.core.contents.sections.views import SectionView
+from imio.smartweb.core.contents.sections.views import HashableJsonSectionView
 from imio.smartweb.core.utils import batch_results
 from imio.smartweb.core.utils import get_json
 from imio.smartweb.core.utils import hash_md5
@@ -18,7 +18,7 @@ from zope.i18nmessageid import MessageFactory
 import json
 
 
-class ContactView(SectionView):
+class ContactView(HashableJsonSectionView):
     """Contact Section view"""
 
     @property
@@ -30,13 +30,11 @@ class ContactView(SectionView):
         current_lang = api.portal.get_current_language()[:2]
         if current_lang != "fr":
             url = f"{url}&translated_in_{current_lang}=1"
-        json_search_contact = get_json(url)
-        if (
-            json_search_contact is None
-            or len(json_search_contact.get("items", [])) == 0  # NOQA
-        ):
+        self.json_data = get_json(url)
+        self.refresh_modification_date()
+        if self.json_data is None or len(self.json_data.get("items", [])) == 0:  # NOQA
             return
-        return json_search_contact.get("items")[0]
+        return self.json_data.get("items")[0]
 
     @property
     def contact_type_class(self):