+++ src/imio/smartweb/core/contents/sections/events/view.py
@@ -4,6 +4,7 @@ from datetime import date
 from dateutil.parser import parse
 from imio.smartweb.core.config import EVENTS_URL
 from imio.smartweb.core.contents.sections.views import CarouselOrTableSectionView
+from imio.smartweb.core.contents.sections.views import HashableJsonSectionView
 from imio.smartweb.core.utils import batch_results
 from imio.smartweb.core.utils import get_json
 from imio.smartweb.core.utils import hash_md5
@@ -11,7 +12,7 @@ from plone import api
 from Products.CMFPlone.utils import normalizeString
 
 
-class EventsView(CarouselOrTableSectionView):
+class EventsView(CarouselOrTableSectionView, HashableJsonSectionView):
     """Events Section view"""
 
     @property
@@ -44,15 +45,13 @@ class EventsView(CarouselOrTableSectionView):
                 "sort_on=event_dates",
             ]
         url = "{}/@events?{}".format(EVENTS_URL, "&".join(params))
-        json_search_events = get_json(url)
-        if (
-            json_search_events is None
-            or len(json_search_events.get("items", [])) == 0  # NOQA
-        ):
+        self.json_data = get_json(url)
+        self.refresh_modification_date()
+        if self.json_data is None or len(self.json_data.get("items", [])) == 0:  # NOQA
             return []
         linking_view_url = self.context.linking_rest_view.to_object.absolute_url()
         image_scale = self.image_scale
-        items = json_search_events.get("items")[:max_items]
+        items = self.json_data.get("items")[:max_items]
         results = []
         for item in items:
             item_id = normalizeString(item["title"])