+++ src/imio/smartweb/core/contents/sections/news/view.py
@@ -2,6 +2,7 @@
 
 from imio.smartweb.core.config import NEWS_URL
 from imio.smartweb.core.contents.sections.views import CarouselOrTableSectionView
+from imio.smartweb.core.contents.sections.views import HashableJsonSectionView
 from imio.smartweb.core.utils import batch_results
 from imio.smartweb.core.utils import get_json
 from imio.smartweb.core.utils import hash_md5
@@ -9,7 +10,7 @@ from plone import api
 from Products.CMFPlone.utils import normalizeString
 
 
-class NewsView(CarouselOrTableSectionView):
+class NewsView(CarouselOrTableSectionView, HashableJsonSectionView):
     """News Section view"""
 
     @property
@@ -40,15 +41,13 @@ class NewsView(CarouselOrTableSectionView):
                 "sort_order=descending",
             ]
         url = "{}/@search?{}".format(NEWS_URL, "&".join(params))
-        json_search_news = get_json(url)
-        if (
-            json_search_news is None
-            or len(json_search_news.get("items", [])) == 0  # NOQA
-        ):
+        self.json_data = get_json(url)
+        self.refresh_modification_date()
+        if self.json_data is None or len(self.json_data.get("items", [])) == 0:  # NOQA
             return []
         linking_view_url = self.context.linking_rest_view.to_object.absolute_url()
         image_scale = self.image_scale
-        items = json_search_news.get("items")[:max_items]
+        items = self.json_data.get("items")[:max_items]
         results = []
         for item in items:
             item_id = normalizeString(item["title"])