+++ CHANGES.rst
@@ -5,8 +5,12 @@ Changelog
 1.1.15 (unreleased)
 -------------------
 
+- Store hash in an annotation to refresh "dynamic" sections
+  [boulch, laulaz]
+
 - WEB-3868 : Remove useless code (included in Plone 6.0.4)
   See https://github.com/plone/plone.base/pull/37
+  [laulaz]
 
 - Migrate to Plone 6.0.4
   [boulch]